# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AspApi",
    "version": "1.0"
  },
  "paths": {
...
    "/api/user/getuserswithpagination": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "draw",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "filter",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/userexists": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "userName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/emailexists": {
      "get": {
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/usernameoremailexists": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UsernameOrEmailExistsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UsernameOrEmailExistsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UsernameOrEmailExistsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/user/addnewuser": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "$ref": "#/components/schemas/DateOnly"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "address2": {
                    "type": "string"
                  },
                  "province": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "zipCode": {
                    "type": "string"
                  },
                  "userRole": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "avatarFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "userName": {
                  "style": "form"
                },
                "email": {
                  "style": "form"
                },
                "password": {
                  "style": "form"
                },
                "dateOfBirth": {
                  "style": "form"
                },
                "firstName": {
                  "style": "form"
                },
                "lastName": {
                  "style": "form"
                },
                "address": {
                  "style": "form"
                },
                "address2": {
                  "style": "form"
                },
                "province": {
                  "style": "form"
                },
                "city": {
                  "style": "form"
                },
                "zipCode": {
                  "style": "form"
                },
                "userRole": {
                  "style": "form"
                },
                "isActive": {
                  "style": "form"
                },
                "avatarFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/updateuser": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "$ref": "#/components/schemas/DateOnly"
                  },
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "address2": {
                    "type": "string"
                  },
                  "province": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "zipCode": {
                    "type": "string"
                  },
                  "userRole": {
                    "type": "string"
                  },
                  "isActive": {
                    "type": "integer",
                    "format": "int32"
                  },
                  "avatarFile": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "userName": {
                  "style": "form"
                },
                "email": {
                  "style": "form"
                },
                "password": {
                  "style": "form"
                },
                "dateOfBirth": {
                  "style": "form"
                },
                "firstName": {
                  "style": "form"
                },
                "lastName": {
                  "style": "form"
                },
                "address": {
                  "style": "form"
                },
                "address2": {
                  "style": "form"
                },
                "province": {
                  "style": "form"
                },
                "city": {
                  "style": "form"
                },
                "zipCode": {
                  "style": "form"
                },
                "userRole": {
                  "style": "form"
                },
                "isActive": {
                  "style": "form"
                },
                "avatarFile": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/deleteuser": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserResetPassword"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UserResetPassword"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UserResetPassword"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/changepassword": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "userName": {
                  "style": "form"
                },
                "oldPassword": {
                  "style": "form"
                },
                "newPassword": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/user/overidepassword": {
      "post": {
        "tags": [
          "User"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userName": {
                    "type": "string"
                  },
                  "oldPassword": {
                    "type": "string"
                  },
                  "newPassword": {
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "userName": {
                  "style": "form"
                },
                "oldPassword": {
                  "style": "form"
                },
                "newPassword": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "DateOnly": {
        "type": "object",
        "properties": {
          "year": {
            "type": "integer",
            "format": "int32"
          },
          "month": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "type": "integer",
            "format": "int32"
          },
          "dayOfWeek": {
            "$ref": "#/components/schemas/DayOfWeek"
          },
          "dayOfYear": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          },
          "dayNumber": {
            "type": "integer",
            "format": "int32",
            "readOnly": true
          }
        },
        "additionalProperties": false
      },
      "DayOfWeek": {
        "enum": [0, 1, 2, 3, 4, 5, 6],
        "type": "integer",
        "format": "int32"
      },
...
      "UserResetPassword": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "oldPassword": {
            "type": "string",
            "nullable": true
          },
          "newPassword": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "UsernameOrEmailExistsRequest": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "password": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}